refector: RockPaperScissorsGame class, enum RPS 파일 분리
