+++ RockPaperScissors/RockPaperScissors/RPS.swift
@@ -1,0 +1,13 @@
+//
+//  RPS.swift
+//  RockPaperScissors
+//
+//  Created by 정정수 on 2022/08/22.
+//
+
+
+enum RPS: Int {
+    case scissors = 1
+    case rock = 2
+    case paper = 3
+}
